+++ packages/strapi-connector-mongoose/lib/queries.js
@@ -4,11 +4,9 @@
  */
 
 const _ = require('lodash');
-var semver = require('semver');
 const { convertRestQueryParams, buildQuery } = require('strapi-utils');
 
 const { findComponentByGlobalId } = require('./utils/helpers');
-const utils = require('./utils')();
 
 const hasPK = (obj, model) => _.has(obj, model.primaryKey) || _.has(obj, 'id');
 const getPK = (obj, model) => (_.has(obj, model.primaryKey) ? obj[model.primaryKey] : obj.id);
@@ -484,27 +482,23 @@ module.exports = ({ model, strapi }) => {
   function search(params, populate) {
     const populateOpt = populate || defaultPopulate;
 
-    const search = buildSearchOr(model, params._q);
-    _.unset(params, '_q');
-
-    const filters = convertRestQueryParams(params);
-
-    return buildQuery({
-      model,
-      filters,
-      search,
-      populate: populateOpt,
-    }).then(results => results.map(result => (result ? result.toObject() : null)));
-  }
-
-  function countSearch(params) {
-    const search = buildSearchOr(model, params._q);
     const filters = convertRestQueryParams(_.omit(params, '_q'));
 
     return buildQuery({
       model,
       filters,
-      search,
+      searchParam: params._q,
+      populate: populateOpt,
+    }).then(results => results.map(result => (result ? result.toObject() : null)));
+  }
+
+  function countSearch(params) {
+    const filters = convertRestQueryParams(_.omit(params, '_q'));
+
+    return buildQuery({
+      model,
+      filters,
+      searchParam: params._q,
     }).count();
   }
 
@@ -518,48 +512,6 @@ module.exports = ({ model, strapi }) => {
     search,
     countSearch,
   };
-};
-
-const buildSearchOr = (model, query) => {
-  const searchOr = Object.keys(model.attributes).reduce((acc, curr) => {
-    switch (model.attributes[curr].type) {
-      case 'biginteger':
-      case 'integer':
-      case 'float':
-      case 'decimal':
-        if (!_.isNaN(_.toNumber(query))) {
-          const mongoVersion = model.db.base.mongoDBVersion;
-          if (semver.valid(mongoVersion) && semver.gt(mongoVersion, '4.2.0')) {
-            return acc.concat({
-              $expr: {
-                $regexMatch: {
-                  input: { $toString: `$${curr}` },
-                  regex: _.escapeRegExp(query),
-                },
-              },
-            });
-          } else {
-            return acc.concat({ [curr]: _.toNumber(query) });
-          }
-        }
-        return acc;
-      case 'string':
-      case 'text':
-      case 'richtext':
-      case 'email':
-      case 'enumeration':
-      case 'uid':
-        return acc.concat({ [curr]: { $regex: _.escapeRegExp(query), $options: 'i' } });
-      default:
-        return acc;
-    }
-  }, []);
-
-  if (utils.isMongoId(query)) {
-    searchOr.push({ _id: query });
-  }
-
-  return searchOr;
 };
 
 function validateRepeatableInput(value, { key, min, max, required }) {
